[level 1] Title: 숫자 문자열과 영단어, Time: 0.01 ms, Memory: 4.2 MB -BaekjoonHub

+++ 프로그래머스/1/81301. 숫자 문자열과 영단어/숫자 문자열과 영단어.cpp
@@ -83,7 +83,7 @@ for(int i=0;i<s.size();++i)
 int a = 0;
 for (int i = 0; i < c.length(); ++i)
 {
-	a = a * 10 + static_cast<int>(c[i]-48);
+	a = a * 10 + (c[i]-48);
 }
     return a;
 }
